requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
opencv-python
google-generativeai
python-dotenv
